Final Build Until Youtube MP3 Solution
This will be the final build available for an unknown amount of time. Youtube has patched MP3 conversions of their videos, making this program unusable.

In this build, I implemented log in procedures with Spotify's API, including valid expiration of its tokens. This version also shows all user playlists in a miniwindow, each with the ability to be selected via square box. I finally implemented a settings window with log out and change download folder features.